requirements.txt
fastapi
uvicorn[standard]
redis
rq
requests
python-dotenv
pydantic
